flask-backend/requirements.txt
Flask==2.0.1
tensorflow==2.5.0
google-cloud-storage==1.42.3